+++ commandLineApp/commandLineApp.node.ts
@@ -158,7 +158,7 @@ ${errors.length} errors found ${errors.length ? "\n" + errors.join("\n") : ""}`
 
   _check(programPath: treeNotationTypes.treeProgramFilePath) {
     const grammarPath = this._getGrammarPathOrThrow(programPath)
-    const program = jtree.makeProgram(programPath, grammarPath)
+    const program = jtree.compileGrammarAndCreateProgram(programPath, grammarPath)
     return program.getAllErrors().map((err: any) => err.getMessage())
   }
 
@@ -213,7 +213,7 @@ ${errors.length} errors found ${errors.length ? "\n" + errors.join("\n") : ""}`
   compile(programPath: treeNotationTypes.treeProgramFilePath) {
     // todo: allow user to provide destination
     const grammarPath = this._getGrammarPathOrThrow(programPath)
-    const program = jtree.makeProgram(programPath, grammarPath)
+    const program = jtree.compileGrammarAndCreateProgram(programPath, grammarPath)
     const grammarProgram = new HandGrammarProgram(Disk.read(grammarPath))
     return program.compile()
   }
@@ -305,7 +305,7 @@ ${errors.length} errors found ${errors.length ? "\n" + errors.join("\n") : ""}`
       const result = await program.execute()
       return result
     }
-    const result = await jtree.executeFile(programPath, grammarPath)
+    const result = jtree.compileGrammarAndCreateProgram(programPath, grammarPath).execute(programPath)
     return result
   }