+++ coreTests/factory.node.test.ts
@@ -5,14 +5,14 @@ import { treeNotationTypes } from "../products/treeNotationTypes"
 
 const testTree: treeNotationTypes.testTree = {}
 
-testTree.makeProgram = equal => {
+testTree.compileGrammarAndCreateProgram = equal => {
   // Arrange
   const jibberishRootDir = __dirname + "/../langs/jibberish/"
   const programPath = jibberishRootDir + "sample.jibberish"
   const grammarPath = jibberishRootDir + "jibberish.grammar"
 
   // Act
-  const program = jtree.makeProgram(programPath, grammarPath)
+  const program = jtree.compileGrammarAndCreateProgram(programPath, grammarPath)
   const result = program.execute()
 
   // Assert