+++ releaseNotes.md
@@ -8,9 +8,10 @@ edge / 2019-12-08
 - Breaking: removed jtree.executeFiles
 - Breaking: removed all /langs/lang/lang.node.js files. Use /products/lang.nodejs.js instead.
 - Breaking: removed commandLineApp "runSync" method.
-- Breaking: removed jtree.executeFileSync method
+- Breaking: removed jtree.executeFileSync and executeFile methods
 - Breaking: removed executeSync method on Core. Up to the grammar to define it's own execute methods.
 - Breaking: removed default execute implementation on Core. Up to grammar to specify.
+- Breaking: jtree.makeProgram is now jtree.compileGrammarAndCreateProgram
 
 48.1.0 / 2019-12-06
 ===================